리팩토링: Traffic클래스로 분리
